back working and front still needs work

+++ client/src/services/api-helper.jsx
@@ -7,50 +7,50 @@ const api = axios.create({
 
 export const loginUser = async (loginData) => {
   const resp = await api.post('/auth/login', loginData)
-  console.log(resp)
+  // console.log(resp)
   localStorage.setItem('authToken', resp.data.token);
   api.defaults.headers.common.authorization = `Bearer ${resp.data.token}`
   return resp.data.user
 }
 
 export const registerUser = async (registerData) => {
-  const resp = await api.post('/users/', { user: registerData })
+  const resp = await api.post('/users', { user: registerData })
   return resp.data
 }
 
-export const verifyUser = async () => {
-  const token = localStorage.getItem('authToken');
-  if (token) {
-    api.defaults.headers.common.authorization = `Bearer ${token}`
-    const resp = await api.get('/users/verify');
-    return resp.data
-  }
-  return false;
-}
+// export const verifyUser = async () => {
+//   const token = localStorage.getItem('authToken');
+//   if (token) {
+//     api.defaults.headers.common.authorization = `Bearer ${token}`
+//     const resp = await api.get('/users/verify');
+//     return resp.data
+//   }
+//   return false;
+// }
 
-const createMatches = async (data) => {
-  const resp = await api.post('/matches', { matches: data })
+const createMatches = async (data, user_id) => {
+  const resp = await api.post(`/users/${user_id}/matches`, { matches: data })
   return resp.data
 }
 
-const readAllMatches = async () => {
-  const resp = await api.get('/matches')
+const readAllMatches = async (user_id) => {
+  const resp = await api.get(`/users/${user_id}/matches`)
   return resp.data
 }
 
-const readOneMatches = async (id) => {
-  const resp = await api.get(`/matches/${id}`)
+const readOneMatches = async (id, user_id) => {
+  const resp = await api.get(`/users/${user_id}/matches/${id}`)
   return resp.data
 
 }
 
-const updateMatches = async (id, data) => {
-  const resp = await api.put(`/matches/${id}`, { matches: data })
+const updateMatches = async (id, data, user_id) => {
+  const resp = await api.put(`/users/${user_id}/matches/${id}`, { matches: data })
   return resp.data
 }
 
-const destroyMatches = async (id) => {
-  const resp = await api.delete(`/matches/${id}`)
+const destroyMatches = async (id, user_id) => {
+  const resp = await api.delete(`/users/${user_id}/matches/${id}`)
   return resp.data
 
 }
